Controller inputs, Dash, Grab, Dash Animations, Update Assets

+++ Celeste/MusicManager.cpp
@@ -66,6 +66,6 @@ void MusicManager::MuteVolume()
 
 void MusicManager::UnmuteVolume(int _sound)
 {
-	_musicData->setVolume(_sound);
+	_musicData->setVolume(static_cast<float>(_sound));
 }
